事务

+++ src/main/java/com/bdqn/dao/BookDao.java
@@ -6,5 +6,5 @@ public interface BookDao {
     //更新书的库存
     void updateBookStock(int bid);
     //更新用户的账户余额
-    void updateUserAccount(String usename,int price);
+    void updateUserAccount(String usename,double price);
 }
